+++ app/controllers/jobs.controller.js
@@ -5,7 +5,19 @@
  * @module JobsController
  */
 
+const ExportService = require('../services/jobs/export/export.service.js')
 const ProcessTimeLimitedLicencesService = require('../services/jobs/time-limited/process-time-limited-licences.service.js')
+
+/**
+ * Triggers export of all relevant tables to CSV and then uploads them to S3
+ *
+ * > Has to be called something other than 'export' because export is a reserved word
+ */
+async function exportDb (_request, h) {
+  ExportService.go()
+
+  return h.response().code(204)
+}
 
 async function timeLimited (request, h) {
   ProcessTimeLimitedLicencesService.go()
@@ -14,5 +26,6 @@ async function timeLimited (request, h) {
 }
 
 module.exports = {
+  exportDb,
   timeLimited
 }
